+++ include/Layer/VGGLayer.hpp
@@ -19,10 +19,10 @@
 #include "Layer/Core/TransformNode.hpp"
 #include "Layer/Core/ZoomerNode.hpp"
 #include "Layer/Core/PaintNode.hpp"
+#include "Layer/Core/ViewportNode.hpp"
 #include "Layer/Config.hpp"
 #include "Layer/Effects.hpp"
 #include "Layer/Graphics/GraphicsLayer.hpp"
-#include "Layer/ViewportNode.hpp"
 
 #include <vector>
 class SkCanvas;
